api/app/requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.5.0
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hashids==1.3.1
httptools==0.2.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2==2.9.7
pyasn1==0.5.0
pydantic==1.10.4
pydantic_core==2.6.3
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.16.0
watchgod==0.7
websockets==10.1
